refactor: reactions use the serenity object as a store now

+++ src/chat/src/reaction.rs
@@ -13,7 +13,7 @@ pub enum ReactionEmoji {
 }
 
 pub trait MessageReaction {
-  fn get_count(&self, kind: Option<MessageReactionType>) -> u32;
+  fn get_count(&self, kind: Option<MessageReactionType>) -> u64;
   fn get_self_reaction(&self) -> Option<MessageReactionType>;
   fn get_emoji(&self) -> ReactionEmoji;
   fn get_burst_colors(&self) -> Vec<Rgba>;
